+++ index.php
@@ -20,6 +20,7 @@ require_once dirname(CONNECTOR_DIR) . '/GXMainComponents/Application.inc.php';
 if (!strpos($_SERVER['REQUEST_URI'], 'jtlconnector/install')) {
     $connector = Connector::getInstance();
     $application = Application::getInstance();
+    $application->createFeaturesFileIfNecessary(sprintf('%s/config/features.json.example', CONNECTOR_DIR));
     $application->register($connector);
     $application->run();
 }