+++ src/Connector.php
@@ -3,18 +3,17 @@
 namespace jtl\Connector\Gambio;
 
 use jtl\Connector\Core\Exception\DatabaseException;
+use jtl\Connector\Core\Rpc\Error;
 use \jtl\Connector\Core\Rpc\RequestPacket;
 use \jtl\Connector\Core\Utilities\RpcMethod;
 use \jtl\Connector\Core\Database\Mysql;
 use jtl\Connector\Gambio\Controller\BaseController;
 use jtl\Connector\Gambio\Gambio\Application;
+use jtl\Connector\Event\Connector\ConnectorAfterFinishEvent;
 use jtl\Connector\Gambio\Util\ConfigHelper;
 use jtl\Connector\Gambio\Util\ShopVersion;
-use jtl\Connector\Model\DeliveryNote;
-use jtl\Connector\Model\Product;
-use jtl\Connector\Model\ProductPrice;
-use jtl\Connector\Model\ProductStockLevel;
-use jtl\Connector\Model\StatusChange;
+use jtl\Connector\Model\DataModel;
+use jtl\Connector\Model\Image;
 use \jtl\Connector\Session\SessionHelper;
 use \jtl\Connector\Base\Connector as BaseConnector;
 use \jtl\Connector\Core\Rpc\Method;
@@ -25,6 +24,12 @@ use \jtl\Connector\Gambio\Checksum\ChecksumLoader;
 
 class Connector extends BaseConnector
 {
+    /**
+     * @var BaseController
+     */
+    public const
+        FINISH_TASK_CLEANUP_PRODUCT_PROPERTIES = 'cleanup_product_properties';
+
     /**
      * @var Application
      */
@@ -82,9 +87,29 @@ class Connector extends BaseConnector
         $this->setPrimaryKeyMapper(new PrimaryKeyMapper());
         $this->setTokenLoader(new TokenLoader());
         $this->setChecksumLoader(new ChecksumLoader());
+
+        $this->getEventDispatcher()->addListener(ConnectorAfterFinishEvent::EVENT_NAME, function(ConnectorAfterFinishEvent $event) use($db) {
+            if(isset($_SESSION[self::FINISH_TASK_CLEANUP_PRODUCT_PROPERTIES]) && $_SESSION[self::FINISH_TASK_CLEANUP_PRODUCT_PROPERTIES] === true) {
+                $queries = [
+                    'DELETE pv FROM properties_values pv WHERE pv.properties_values_id NOT IN (SELECT properties_values_id FROM products_properties_combis_values);',
+                    'DELETE pvd FROM properties_values_description pvd WHERE pvd.properties_values_id NOT IN (SELECT properties_values_id FROM properties_values);',
+                    'DELETE p FROM properties p WHERE p.properties_id NOT IN (SELECT properties_id FROM properties_values);',
+                    'DELETE pd FROM properties_description pd WHERE pd.properties_id NOT IN (SELECT properties_id FROM properties);'
+                ];
+
+                foreach($queries as $sql) {
+                    $db->query($sql);
+                }
+
+                $_SESSION[self::FINISH_TASK_CLEANUP_PRODUCT_PROPERTIES] = false;
+            }
+        });
     }
 
-    private function update($db)
+    /**
+     * @param Mysql $db
+     */
+    protected function update(Mysql $db): void
     {
         if (version_compare(file_get_contents(CONNECTOR_DIR.'/db/version'), CONNECTOR_VERSION) == -1) {
             $versions = [];
@@ -94,11 +119,12 @@ class Connector extends BaseConnector
                 }
             }
 
-            sort($versions);
+            usort($versions, 'version_compare');
 
             foreach ($versions as $version) {
                 if (version_compare(file_get_contents(CONNECTOR_DIR.'/db/version'), $version) == -1) {
                     include(CONNECTOR_DIR.'/db/updates/' . $version . '.php');
+                    file_put_contents(CONNECTOR_DIR.'/db/version', $version);
                 }
             }
         }
@@ -127,7 +153,7 @@ class Connector extends BaseConnector
 
         if ($this->action === Method::ACTION_PUSH || $this->action === Method::ACTION_DELETE) {
             if (!is_array($requestpacket->getParams())) {
-                throw new \Exception('data is not an array');
+                throw new \Exception('Data is not an array');
             }
 
             $action = new Action();
@@ -135,27 +161,19 @@ class Connector extends BaseConnector
 
             $link = Mysql::getInstance();
             $link->DB()->begin_transaction();
-            
-            foreach ($requestpacket->getParams() as $param) {
-                $result = $this->controller->{$this->action}($param);
 
-                $reflectionClass = new \ReflectionClass($param);
-    
+            /** @var DataModel $model */
+            foreach ($requestpacket->getParams() as $model) {
+                $result = $this->controller->{$this->action}($model);
+
                 if ($result->getError()) {
+
                     $link->rollback();
-                    $message = sprintf('Type: %s %s', get_class($param), $result->getError()->getMessage());
-                    
-                    if ($param instanceof Product) {
-                        $message = sprintf('Type: Product Host-Id: %s SKU: %s %s', $param->getId()->getHost(), $param->getSku(), $result->getError()->getMessage());
-                    } elseif ($param instanceof ProductPrice || $param instanceof ProductStockLevel) {
-                        $message = sprintf('Type: %s Product Host-Id: %s %s', $reflectionClass->getShortName(), $param->getProductId()->getHost(), $result->getError()->getMessage());
-                    } elseif ($param instanceof StatusChange || $param instanceof DeliveryNote) {
-                        $message = sprintf('Type: %s Order Host-Id: %s %s', $reflectionClass->getShortName(), $param->getCustomerOrderId()->getHost(), $result->getError()->getMessage());
-                    } elseif (method_exists($param, 'getId')) {
-                        $message = sprintf('Type: %s Host-Id: %s %s', $reflectionClass->getShortName(), $param->getId()->getHost(), $result->getError()->getMessage());
+
+                    if ($result->getError()) {
+                        $this->extendErrorMessage($model, $result->getError());
+                        throw new \Exception($result->getError()->getMessage());
                     }
-                    
-                    throw new \Exception($message);
                 }
                 
                 $results[] = $result->getResult();
@@ -170,6 +188,42 @@ class Connector extends BaseConnector
             return $action;
         } else {
             return $this->controller->{$this->action}($requestpacket->getParams());
+        }
+    }
+
+    /**
+     * @param DataModel $model
+     * @param Error $error
+     */
+    protected function extendErrorMessage(DataModel $model, Error $error)
+    {
+        $controllerToIdentityGetter = [
+            'ProductPrice' => 'getProductId',
+            'ProductStockLevel' => 'getProductId',
+            'StatusChange' => 'getCustomerOrderId',
+            'DeliveryNote' => 'getCustomerOrderId',
+            'Image' => 'getForeignKey',
+        ];
+
+        $controllerName = (new \ReflectionClass($this->controller))->getShortName();
+
+        $identityGetter = $controllerToIdentityGetter[$controllerName] ?? 'getId';
+        $identity = null;
+        if (method_exists($model, $identityGetter)) {
+            $identity = $model->{$identityGetter}();
+        }
+
+        if ($identity !== null) {
+            $messageParts = [$controllerName];
+
+            if ($model instanceof Image) {
+                $messageParts[] = sprintf('Related type %s (hostId = %d)', ucfirst($model->getRelationType()), $identity->getHost());
+            } else {
+                $messageParts[] = sprintf('hostId = %d', $identity->getHost());
+            }
+
+            $messageParts[] = $error->getMessage();
+            $error->setMessage(implode(' | ', $messageParts));
         }
     }