+++ src/Installer/Installer.php
@@ -24,97 +24,110 @@ class Installer
      * @var ConfigHelper
      */
     protected $configHelper;
+
+    protected $db;
+
+    protected $shopConfig;
     
     public function __construct()
     {
         error_reporting(E_ALL ^ E_NOTICE);
         ini_set('display_errors', 1);
         
-        $db = Mysql::getInstance();
-        $this->configHelper = new ConfigHelper($db);
+        $this->db = Mysql::getInstance();
+        $this->configHelper = new ConfigHelper($this->db);
 
-        $shopConfig = $this->configHelper->readGxConfigFile();
+        $this->shopConfig = $this->configHelper->readGxConfigFile();
         $this->connectorConfig = new Config(CONNECTOR_DIR . '/config/config.json');
 
 
-        if (!$db->isConnected()) {
-            $db->connect([
-                "host"     => $shopConfig['db']["host"],
-                "user"     => $shopConfig['db']["user"],
-                "password" => $shopConfig['db']["pass"],
-                "name"     => $shopConfig['db']["name"],
+        if (!$this->db->isConnected()) {
+            $this->db->connect([
+                "host"     => $this->shopConfig['db']["host"],
+                "user"     => $this->shopConfig['db']["user"],
+                "password" => $this->shopConfig['db']["pass"],
+                "name"     => $this->shopConfig['db']["name"],
             ]);
         }
         
-        $db->setNames();
-        
+        $this->db->setNames();
+    }
+
+    /**
+     * @return string
+     */
+    public function runAndGetData(): string
+    {
         $moduleInstances = [];
-        
+
+        $html = '';
         foreach ($this->modules as $id => $module) {
             $className = '\\jtl\\Connector\\Gambio\\Installer\\Modules\\' . $module;
-            $moduleInstances[$id] = new $className($db, $this->connectorConfig, $this->configHelper, $shopConfig);
+            $moduleInstances[$id] = new $className($this->db, $this->connectorConfig, $this->configHelper, $this->shopConfig);
         }
-        
+
         if ($moduleInstances['check']->hasPassed()) {
-            echo '<ul class="nav nav-tabs">';
-            
+            $html .= '<ul class="nav nav-tabs">';
+
             foreach ($moduleInstances as $class => $instance) {
                 $active = $class == 'check' ? 'active' : '';
-                echo '<li class="' . $active . '"><a href="#' . $class . '" data-toggle="tab"><b>' . $instance::$name . '</b></a></li>';
+                $html .= '<li class="' . $active . '"><a href="#' . $class . '" data-toggle="tab"><b>' . $instance::$name . '</b></a></li>';
             }
-            
-            echo '</ul>
+
+            $html .= '</ul>
 	        	<br>
 	        	<div class="tab-content">';
-            
+
             $moduleErrors = [];
-            
+
             foreach ($moduleInstances as $class => $instance) {
                 $active = $class == 'check' ? ' active' : '';
-                
+
                 if (isset($_REQUEST['save'])) {
                     $moduleSave = $instance->save();
                     if ($moduleSave !== true) {
                         $moduleErrors[] = $moduleSave;
                     }
                 }
-                
-                echo '<div class="tab-pane' . $active . '" id="' . $class . '">';
-                echo $instance->form();
-                echo '</div>';
+
+                $html .= '<div class="tab-pane' . $active . '" id="' . $class . '">';
+                $html .= $instance->form();
+                $html .= '</div>';
             }
-            
-            echo '</div>';
-            
+
+            $html .= '</div>';
+
             if (isset($_REQUEST['save'])) {
                 if (count($moduleErrors) == 0) {
                     if (!$this->connectorConfig->save()) {
-                        echo '<div class="alert alert-danger">Fehler beim Schreiben der config.json Datei.</div>';
+                        $html .= '<div class="alert alert-danger">Fehler beim Schreiben der config.json Datei.</div>';
                     } else {
-                        echo '<div class="alert alert-success">Connector Konfiguration wurde gespeichert.</div>';
-                        echo '<div class="alert alert-danger"><b>ACHTUNG:</b><br/>
+                        $html .= '<div class="alert alert-success">Connector Konfiguration wurde gespeichert.</div>';
+                        $html .= '<div class="alert alert-danger"><b>ACHTUNG:</b><br/>
                             Bitte sorgen Sie nach erfolgreicher Installation des Connectors unbedingt dafür, dass dieser Installer 
                             sowie die Datei config.json im Verzeichnis config nicht öffentlich les- und ausführbar sind!</div>';
                     }
                 } else {
-                    echo '<div class="alert alert-danger">Folgende Fehler traten auf:
+                    $html .= '<div class="alert alert-danger">Folgende Fehler traten auf:
 		        		<br>
 		        		<ul>';
-                    
+
                     foreach ($moduleErrors as $error) {
-                        echo '<li>' . $error . '</li>';
+                        $html .= '<li>' . $error . '</li>';
                     }
-                    
-                    echo '</ul>
+
+                    $html .= '</ul>
 		        		</div>';
                 }
             }
-            
-            echo '<button type="submit" name="save" class="btn btn-primary btn-block"><span class="glyphicon glyphicon-save"></span> Connector Konfiguration speichern</button>';
+
+            $html .= '<button type="submit" name="save" class="btn btn-primary btn-block"><span class="glyphicon glyphicon-save"></span> Connector Konfiguration speichern</button>';
         } else {
-            echo '<div class="alert alert-danger">Bitte beheben Sie die angezeigten Fehler bevor Sie mit der Konfiguration fortfahren können.</div>';
-            echo $moduleInstances['check']->form();
+            $html .= '<div class="alert alert-danger">Bitte beheben Sie die angezeigten Fehler bevor Sie mit der Konfiguration fortfahren können.</div>';
+            $html .= $moduleInstances['check']->form();
         }
+
+        return $html;
     }
     
     private function readConfigFile()